requirements-phase2.txt
# Congressional Trading Intelligence System - Phase 2 Requirements
# Intelligence & Analytics Dependencies

# Inherit Phase 1 requirements
-r requirements-phase1.txt

# Machine Learning and Data Science
scikit-learn>=1.2.0
tensorflow>=2.11.0
torch>=1.13.0
xgboost>=1.7.0
lightgbm>=3.3.0
joblib>=1.2.0
pickle5>=0.0.12

# Advanced Analytics
scipy>=1.10.0
statsmodels>=0.13.0
networkx>=3.0
community>=1.0.0
node2vec>=0.4.0

# Natural Language Processing
transformers>=4.25.0
torch-audio>=2.0.0
spacy>=3.4.0
nltk>=3.8.0
textblob>=0.17.0
vaderSentiment>=3.3.0
newspaper3k>=0.2.8

# Time Series Analysis
prophet>=1.1.0
tslearn>=0.5.0
arch>=5.3.0

# Visualization Libraries
plotly>=5.12.0
bokeh>=3.0.0
altair>=4.2.0
seaborn>=0.12.0
matplotlib>=3.6.0
wordcloud>=1.9.0

# Interactive Dashboard (React Migration)
dash>=2.7.0
dash-bootstrap-components>=1.2.0
dash-cytoscape>=0.3.0
plotly-dash>=1.0.0

# Real-time Processing
celery>=5.2.0
redis>=4.4.0
kafka-python>=2.0.0
websockets>=10.4.0
socketio>=5.7.0

# News and Social Media APIs
newsapi-python>=0.2.6
tweepy>=4.12.0
praw>=7.6.0
feedparser>=6.0.0

# Options and Derivatives Analysis
quantlib-python>=1.29.0
mibian>=0.1.3
py_vollib>=1.0.1
greeks>=1.0.0

# Database Extensions
sqlalchemy-utils>=0.39.0
psycopg2-binary>=2.9.5
alembic>=1.8.0

# API Framework Enhancement
flask>=2.2.0
flask-restful>=0.3.9
flask-cors>=3.0.10
flask-jwt-extended>=4.4.0
flask-limiter>=2.6.0
gunicorn>=20.1.0
uvicorn>=0.20.0
fastapi>=0.88.0

# Caching and Performance
redis>=4.4.0
memcached>=1.0.8
python-memcached>=1.59
cachetools>=5.2.0

# Image Processing and Charts
pillow>=9.4.0
wand>=0.6.10
cairosvg>=2.5.0

# Financial Data
yfinance>=0.2.0
alpha-vantage>=2.3.0
quandl>=3.7.0
pandas-datareader>=0.10.0

# Development Tools
jupyter>=1.0.0
jupyterlab>=3.5.0
notebook>=6.5.0
ipywidgets>=8.0.0

# Model Deployment
mlflow>=2.0.0
dvc>=2.40.0
optuna>=3.0.0

# Testing Enhancement
pytest-mock>=3.10.0
pytest-xdist>=3.1.0
pytest-benchmark>=4.0.0
factory-boy>=3.2.0

# Documentation
sphinx>=5.3.0
sphinx-rtd-theme>=1.1.0
myst-parser>=0.18.0